moved source menu logic to source component
