+++ client/src/components/Other/AddNewSourceComponent.js
@@ -1,9 +1,24 @@
-import React from 'react'
+import React, {useState} from 'react'
 import AddNewSourceMenu from "./AddNewSourceMenu";
 
 const AddNewSourceComponent = () => {
+  const [anchorEl, setAnchorEl] = useState(null);
+  const openMenuBul = Boolean(anchorEl)
+  const openMenu = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+
+  const closeMenu = () => {
+    setAnchorEl(null);
+  }
   return (
-    <AddNewSourceMenu/>
+    <AddNewSourceMenu
+      openMenuFunc={openMenu}
+      openBul={openMenuBul}
+      anchorEl={anchorEl}
+      closeMenuFunc={closeMenu}
+      modalFunc={closeMenu}
+    />
   )
 }
