+++ client/src/components/Other/AddNewSourceModal.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const AddNewSourceModal = () => {
+
+}
+
+export default AddNewSourceModal
